Add to-do list functionality to application
A to-do list module was added to the 'Todos.fs' file to handle the creation, viewing and listing of to-do items. This involved the creation of a 'Views' sub-module to handle the display of an individual to-do item, and a 'Todo' view for rendering the entire list. A hardcoded list of to-do items was subsequently added to the 'Data' sub-module.

In the 'Program.fs', the indexView was updated to accommodate the updated to-do list view. Routes have been updated to route to the new to-do view.

Changes were necessary to provide users with a functional, interactive to-do list. Users can now view a list of to-do items and interact with them accordingly.